expected changes for uninstall

+++ preview/MsixCore/msixmgr/VirtualFileHandler.cpp
@@ -17,6 +17,8 @@ using namespace MsixCoreLib;
 
 const PCWSTR VirtualFileHandler::HandlerName = L"VirtualFileHandler";
 
+std::map < std::wstring, DWORD > m_sharedDllMap;
+
 HRESULT VirtualFileHandler::ExecuteForAddRequest()
 {
     auto vfsDirectoryPath = m_msixRequest->GetPackageDirectoryPath() + L"\\VFS";
@@ -28,7 +30,17 @@ HRESULT VirtualFileHandler::ExecuteForAddRequest()
             RETURN_IF_FAILED(CopyVfsFileToLocal(p.path()));
         }
     }
-    
+
+    // Cases for incrementing. For simplicity, we ignore the count in registry.dat as we treat the .msix package as having one reference to the file
+    // 
+    // File exists, Not in SharedDLLs in registry.dat, already exists in current system SharedDLLs => increment current SharedDLLs + 1
+    // File exists, Not in SharedDLLs in registry.dat, not in current system SharedDLLs => create new SharedDLLs
+    // File exists, exists in SharedDLLs in registry.dat, already exists in current system SharedDLLs => increment current SharedDLLs + 1
+    // File exists, exists in SharedDLLs in registry.dat, not in current system SharedDLLs => create new SharedDLLs(count=1)
+    // File does not exist, does not exist in SharedDLLs in registry.dat, => no SharedDLLs
+    // File does not exist, exists in SharedDLLs in registry.dat => create new SharedDLLs(count=1)
+
+    RETURN_IF_FAILED(m_msixRequest->GetRegistryDevirtualizer()->GetSharedDlls(false));
     
     return S_OK;
 }
@@ -92,6 +104,8 @@ HRESULT VirtualFileHandler::CreateHandler(MsixRequest * msixRequest, IPackageHan
     {
         return E_OUTOFMEMORY;
     }
+    RETURN_IF_FAILED(localInstance->m_sharedDllsKey.Open(HKEY_LOCAL_MACHINE, uninstallKeyPath.c_str(), KEY_WRITE));
+
     *instance = localInstance.release();
 
     return S_OK;
@@ -281,38 +295,85 @@ HRESULT VirtualFileHandler::RemoveVfsFile(std::wstring fileName)
         return S_OK;
     }
 
-    bool success = DeleteFile(fullPath.c_str());
-    if (!success)
+    // Check if file is referenced in SharedDLLs key.
+    // If it is, then just decrement instead of delete; or delete the reg value if it's 0 after decrement
+    UINT32 count = 0;
+    bool sharedDllValueExists = false;
+    bool shouldDelete = true;
+    HRESULT hrGetUInt32Value = m_sharedDllsKey.GetUInt32ValueIfExists(fullPath.c_str(), count, sharedDllValueExists);
+    if (FAILED(hrGetUInt32Value))
     {
         TraceLoggingWrite(g_MsixTraceLoggingProvider,
-            "Unable to Delete file",
+            "Unable to determine if file is shared -- not deleting file",
             TraceLoggingLevel(WINEVENT_LEVEL_WARNING),
             TraceLoggingValue(fullPath.c_str(), "FullPath"),
-            TraceLoggingValue(GetLastError(), "error"));
+            TraceLoggingValue(hrGetUInt32Value, "hr"));
+        shouldDelete = false;
     }
-
-    while (success)
+    else if (sharedDllValueExists)
     {
-        MsixCoreLib_GetPathParent(fullPath);
-
-        // instead of checking if the directory is empty, just try to delete it.
-        // if it's not empty it'll fail with expected error code that we can ignore
-        success = RemoveDirectory(fullPath.c_str());
-        if (!success)
+        if (count > 1)
         {
-            DWORD error = GetLastError();
-            if (error != ERROR_DIR_NOT_EMPTY)
+            shouldDelete = false;
+            HRESULT hrSetUInt32Value = m_sharedDllsKey.SetUInt32Value(fullPath.c_str(), count - 1);
+            if (FAILED(hrSetUInt32Value))
             {
                 TraceLoggingWrite(g_MsixTraceLoggingProvider,
-                    "Unable to Delete directory",
+                    "Unable to decrement sharedDLL key",
                     TraceLoggingLevel(WINEVENT_LEVEL_WARNING),
                     TraceLoggingValue(fullPath.c_str(), "FullPath"),
-                    TraceLoggingValue(GetLastError(), "error"));
+                    TraceLoggingValue(hrSetUInt32Value, "hr"));
             }
         }
-        // if we're successfull in deleting the directory, try to delete the containing directory too, in case it's now empty
+        else
+        {
+            HRESULT hrDeleteValue = m_sharedDllsKey.DeleteValue(fullPath.c_str());
+            if (FAILED(hrDeleteValue))
+            {
+                TraceLoggingWrite(g_MsixTraceLoggingProvider,
+                    "Unable to delete sharedDLL key",
+                    TraceLoggingLevel(WINEVENT_LEVEL_WARNING),
+                    TraceLoggingValue(fullPath.c_str(), "FullPath"),
+                    TraceLoggingValue(hrDeleteValue, "hr"));
+                shouldDelete = false;
+            }
+        }
     }
-    
+
+    if (shouldDelete)
+    {
+        bool success = DeleteFile(fullPath.c_str());
+        if (!success)
+        {
+            TraceLoggingWrite(g_MsixTraceLoggingProvider,
+                "Unable to Delete file",
+                TraceLoggingLevel(WINEVENT_LEVEL_WARNING),
+                TraceLoggingValue(fullPath.c_str(), "FullPath"),
+                TraceLoggingValue(GetLastError(), "error"));
+        }
+
+        while (success)
+        {
+            MsixCoreLib_GetPathParent(fullPath);
+
+            // instead of checking if the directory is empty, just try to delete it.
+            // if it's not empty it'll fail with expected error code that we can ignore
+            success = RemoveDirectory(fullPath.c_str());
+            if (!success)
+            {
+                DWORD error = GetLastError();
+                if (error != ERROR_DIR_NOT_EMPTY)
+                {
+                    TraceLoggingWrite(g_MsixTraceLoggingProvider,
+                        "Unable to Delete directory",
+                        TraceLoggingLevel(WINEVENT_LEVEL_WARNING),
+                        TraceLoggingValue(fullPath.c_str(), "FullPath"),
+                        TraceLoggingValue(GetLastError(), "error"));
+                }
+            }
+            // if we're successfull in deleting the directory, try to delete the containing directory too, in case it's now empty
+        }
+    }
     return S_OK;
 }
